allow UI code and properties to be written in Instruments directly
this allows for flexibilty when designing Instrument classes

also added back Edit Instrument Information (CollectionProperties)

UI elements such as Edit Object Information can now be directly editied in the instrument, unlinking the need to edit the UI code directly

this allows for way more flexibility when designing custom instruments.

I will be working on a feature to detect custom instruments in the Text Editior as well. Custom instruments will be a bit broken for the time being (in the UI portion).

No changes have been made to the functionality to the program, part of the UI is not functional
