+++ MIDIAnimator/src/instruments.py
@@ -1,10 +1,12 @@
 from __future__ import annotations
 
-from typing import Dict, List, Union
+from typing import Dict, List, Union, Type
 from itertools import zip_longest
+from dataclasses import dataclass
+from enum import Enum
 import bpy
 
-from ..data_structures.midi import MIDITrack
+from .. data_structures.midi import MIDITrack
 from .. utils import convertNoteNumbers
 from .. utils.loggerSetup import *
 from .. data_structures import *
@@ -27,13 +29,14 @@ class Instrument:
         """
         self.collection = collection
         self.midiTrack = midiTrack
-        self.noteToWpr = dict()
         self.override = override
+        
+        self.noteToWpr = dict()
         self.preAnimate()
 
         if not self.override:
             self.__post_init__()
-    
+
     def __post_init__(self):
         """checks objects to see if they're valid
         creates note to ObjectWrapper dictionary
@@ -46,6 +49,21 @@ class Instrument:
                 raise ValueError("Animation type `keyframed` must have either a Note On Curve or a Note Off Curve!")
         
         self.createNoteToObjectWpr()
+    
+    @staticmethod
+    def drawInstrument(context: bpy.types.Context, col: bpy.types.UILayout, blCol: bpy.types.Collection,):
+        """draws the UI for the instrument view"""
+        pass
+    
+    @staticmethod
+    def drawObject(context: bpy.types.Context, col: bpy.types.UILayout, blObj: bpy.types.Object):
+        """draws the UI for the object view"""
+        pass
+
+    @staticmethod
+    def properties():
+        """register properties for this instrument"""
+        pass
 
     @staticmethod
     def getFCurves(obj: bpy.types.Object, noteType: str="note_on") -> List[Union[bpy.types.FCurve, ObjectShapeKey]]:
@@ -168,6 +186,33 @@ class ProjectileInstrument(Instrument):
 
         super().__init__(midiTrack, objectCollection)
 
+    @staticmethod
+    def drawInstrument(context: bpy.types.Context, col: bpy.types.UILayout, blCol: bpy.types.Collection,):
+        """draws the UI for the instrument view"""
+        col.prop(blCol.midi, "projectile_collection")
+    
+    @staticmethod
+    def drawObject(context: bpy.types.Context, col: bpy.types.UILayout, blObj: bpy.types.Object):
+        """draws the UI for the object view"""
+        pass
+
+    @staticmethod
+    def properties():
+        """create properties for the instrument"""
+        MIDIAnimatorCollectionProperties.projectile_collection = bpy.props.PointerProperty(
+            name="Projectile Collection",
+            description="the `bpy.types.Collection` to store the projectiles",
+            type=bpy.types.Collection,
+            options=set()
+        )
+
+        MIDIAnimatorCollectionProperties.reference_projectile = bpy.props.PointerProperty(
+            name="Reference Projectile",
+            description="the `bpy.types.Object` to clone the projectile to. This object will not be animated, instead it will be copied.",
+            type=bpy.types.Object,
+            options=set()
+        )
+
     def preAnimate(self):
         """deletes all old projectiles & their associated keyframes"""
         cleanCollection(self.projectileCollection, self.referenceProjectile)
@@ -261,6 +306,125 @@ class EvaluateInstrument(Instrument):
         :param bpy.types.Collection collection: the `bpy.types.Collection` of Blender objects to apply keyframes to
         """
         super().__init__(midiTrack=midiTrack, collection=collection)
+
+    @staticmethod
+    def drawInstrument(context: bpy.types.Context, col: bpy.types.UILayout, blCol: bpy.types.Collection,):
+        """draws the UI for the instrument view"""
+        pass
+    
+    @staticmethod
+    def drawObject(context: bpy.types.Context, col: bpy.types.UILayout, blObj: bpy.types.Object):
+        """draws the UI for the object view"""
+        
+        objMidi = blObj.midi
+
+        col.separator()
+        row0 = col.row()
+
+        row0.prop(objMidi, "anim_type")
+
+        if objMidi.anim_type == "damp_osc":
+            col.prop(objMidi, "osc_period")
+            col.prop(objMidi, "osc_amp")
+            col.prop(objMidi, "osc_damp")
+
+        elif objMidi.anim_type == "keyframed":
+            row1 = col.row()
+            row1.prop(objMidi, "note_on_curve", text="Note On")
+            # this is temporary until I determine a solution for each FCurve
+            row1.prop(objMidi, "note_on_anchor_pt", text="")
+
+            row2 = col.row()
+            row2.prop(objMidi, "note_off_curve", text="Note Off")
+            row2.prop(objMidi, "note_off_anchor_pt", text="")
+
+        elif objMidi.anim_type == "adsr":
+            col.label(text="Coming soon")
+        
+        col.separator()
+
+        col.prop(objMidi, "velocity_intensity", slider=True)
+
+        col.prop(objMidi, "anim_overlap")
+
+    @staticmethod
+    def properties():
+        """create properties for the instrument"""
+        MIDIAnimatorObjectProperties.velocity_intensity = bpy.props.FloatProperty(
+            name="Velocity Intensity", 
+            description="The strength of the velocity affecting the keyframe values. 0 for no velocity influence.",
+            default=0.0,
+            soft_min=0,
+            soft_max=2,
+            options=set()
+        )
+        MIDIAnimatorObjectProperties.note_number_int = bpy.props.IntProperty(
+            name="Note Number Integer", 
+            description="The note number (integer) of an object",
+            default=60
+        )
+        MIDIAnimatorObjectProperties.note_on_curve = bpy.props.PointerProperty(
+            name="Note On Animation Curve", 
+            description="The animation curve object with defined keyframes to be read in",
+            type=bpy.types.Object,
+            options=set()
+        )
+        MIDIAnimatorObjectProperties.note_on_anchor_pt = bpy.props.IntProperty(
+            name="Note On Anchor Point",
+            description="Where along should we start animating (in reference to the note on time). 0 to start animating right on note on times, - for eariler, + for later",
+            default=0,
+            options=set()
+        )
+        MIDIAnimatorObjectProperties.note_off_curve = bpy.props.PointerProperty(
+            name="Note Off Animation Curve", 
+            description="The animation curve object with defined keyframes to be read in."
+                        "\n\nDisabled: will be added in a future release",
+            type=bpy.types.Object,
+            options=set()
+        )
+        MIDIAnimatorObjectProperties.note_off_anchor_pt = bpy.props.IntProperty(
+            name="Note Off Anchor Point",
+            description="Where along should we start animating (in reference to the note on time). 0 to start animating right on note off times, - for eariler, + for later."
+                        "\n\nDisabled: will be added in a future release",
+            default=0,
+            options=set()
+        )
+        MIDIAnimatorObjectProperties.osc_period = bpy.props.FloatProperty(
+            name="Period",
+            description="Period of the oscillation",
+            default=4,
+            options=set()
+        )
+        MIDIAnimatorObjectProperties.osc_amp = bpy.props.FloatProperty(
+            name="Amplitude",
+            description="Amplitude of the oscillation",
+            default=4,
+            options=set()
+        )
+        MIDIAnimatorObjectProperties.osc_damp = bpy.props.FloatProperty(
+            name="Damping",
+            description="Damping of the oscillation",
+            default=10,
+            options=set()
+        )
+        MIDIAnimatorObjectProperties.anim_overlap = bpy.props.EnumProperty(
+            items=[
+                ("add", "Add", "Curves will add motion. More options will be added in the future")
+            ],
+            name="Animation Overlap",
+            default="add",
+            options=set()
+        )
+        MIDIAnimatorObjectProperties.anim_type = bpy.props.EnumProperty(
+            items=[
+                ("keyframed", "Keyframed", "Pre-defined FCurve objects to refernce the animation from"),
+                ("damp_osc", "Oscillation", "Dampened oscillation. Planned for a future release"),
+                ("adsr", "ADSR", "Attack, Decay, Sustain, Release. Planned for a future release")
+            ],
+            name="Animation Type",
+            default="keyframed",
+            options=set()
+        )
 
     def preAnimate(self):
         """cleans all keyframes before the animation"""
@@ -418,3 +582,56 @@ class EvaluateInstrument(Instrument):
                                 fCrv.targetKey.keyframe_insert(data_path="value", frame=keyframe.frame)
                 else:
                     raise RuntimeError(f"ERROR: Type {obj.midi.anim_type} for object {obj.name} not supported yet.")
+
+# ------------------------------------------------------------------
+
+@dataclass
+class InstrumentItem:
+    identifier: str
+    name: str
+    description: str
+    cls: Type[Instrument]
+
+
+# Enum of instruments. Add new instruments here and they will be automatically added to the UI
+class Instruments(Enum):
+    projectile = InstrumentItem(identifier="projectile", name="Projectile", description="A projectile that is fired from the instrument", cls=ProjectileInstrument)
+    evaluate = InstrumentItem(identifier="evaluate", name="Evaluate", description="Evaluate thing", cls=EvaluateInstrument)
+
+
+# Shared properties throughout all objects, instruments(collections) and scences. 
+# For example, a note number is a property likely used by all instruments, so it is defined here
+class MIDIAnimatorObjectProperties(bpy.types.PropertyGroup):
+    note_number:  bpy.props.StringProperty(
+        name="Note Number", 
+        description="The note number of an object. Can be entered as a integer (MIDI Note Number, e.g. 60) or as a "
+                    "readable note (C3)",
+        default="C3"
+    )
+
+class MIDIAnimatorCollectionProperties(bpy.types.PropertyGroup):
+    instrument_type: bpy.props.EnumProperty(
+        items=[(item.value.identifier, item.value.name, item.value.description) for item in Instruments],
+        name="Instrument Type",
+        default="evaluate",
+        options=set()
+    )
+
+class MIDIAnimatorSceneProperties(bpy.types.PropertyGroup):
+    # Edit Notes (Quick)
+    quick_note_number_list: bpy.props.StringProperty(
+        name="Note Number List",
+        description="A list of note numbers. These will correspond to the objects in the selected collection",
+        default="[]",
+        options=set()
+    )
+    quick_obj_col: bpy.props.PointerProperty(
+        type=bpy.types.Collection, 
+        name="Collection",
+    )
+    quick_sort_by_name: bpy.props.BoolProperty(
+        name="Sort by Name",
+        description="This will use a sorted list of objects by name instead of using `name_noteNumber`",
+        default=False,
+        options=set()
+    )
