+++ MIDIAnimator/ui/panels.py
@@ -1,12 +1,45 @@
 from __future__ import annotations
-from .. utils import convertNoteNumbers, typeOfNoteNumber, noteToName
-from ast import literal_eval
+from .. src.instruments import *
 import bpy
 
 class MIDIAniamtorPanel:
     bl_space_type = "VIEW_3D"
     bl_region_type = "UI"
     bl_category = "MIDIAnimator"
+
+class VIEW3D_PT_edit_instrument_information(MIDIAniamtorPanel, bpy.types.Panel):
+    bl_label = "Edit Instrument Information"
+
+    @classmethod
+    def poll(cls, context):
+        selectedObjs = context.selected_editable_objects
+        # return True if collection is selected in outliner
+        return len(selectedObjs) == 0
+
+
+    def draw(self, context):
+        if len(context.selected_editable_objects) != 0:
+            blCol = context.object.users_collection[0]
+        else:
+            # fallback if empty collection
+            blCol = context.collection
+        
+        blColMidi = blCol.midi
+
+        layout = self.layout
+        layout.use_property_decorate = False
+        layout.use_property_split = True
+        col = layout.column()
+        
+        col.label(text=f"Active collection: '{blCol.name}'")
+        col.prop(blColMidi, "instrument_type", text="Instrument Type")
+
+        for item in Instruments:
+            value = item.value
+            if value.identifier == blColMidi.instrument_type:
+                value.cls.drawInstrument(context, col, blCol)
+                break
+
 
 class VIEW3D_PT_edit_object_information(MIDIAniamtorPanel, bpy.types.Panel):
     bl_label = "Edit Object Information"
@@ -20,7 +53,9 @@ class VIEW3D_PT_edit_object_information(MIDIAniamtorPanel, bpy.types.Panel):
     def draw(self, context):
         obj = context.active_object
         objMidi = obj.midi
+
         blCol = obj.users_collection[0]
+        blColMidi = blCol.midi
 
         layout = self.layout
         layout.use_property_decorate = False
@@ -29,36 +64,14 @@ class VIEW3D_PT_edit_object_information(MIDIAniamtorPanel, bpy.types.Panel):
         col = layout.column()
         
         col.label(text=f"Active object: '{obj.name}'")
+
         col.prop(objMidi, "note_number")
         
-        col.separator()
-        row0 = col.row()
-
-        row0.prop(objMidi, "anim_type")
-
-        if objMidi.anim_type == "damp_osc":
-            col.prop(objMidi, "osc_period")
-            col.prop(objMidi, "osc_amp")
-            col.prop(objMidi, "osc_damp")
-
-        elif objMidi.anim_type == "keyframed":
-            row1 = col.row()
-            row1.prop(objMidi, "note_on_curve", text="Note On")
-            # this is temporary until I determine a solution for each FCurve
-            row1.prop(objMidi, "note_on_anchor_pt", text="")
-
-            row2 = col.row()
-            row2.prop(objMidi, "note_off_curve", text="Note Off")
-            row2.prop(objMidi, "note_off_anchor_pt", text="")
-
-        elif objMidi.anim_type == "adsr":
-            col.label(text="Coming soon")
-        
-        col.separator()
-
-        col.prop(objMidi, "velocity_intensity", slider=True)
-
-        col.prop(objMidi, "anim_overlap")
+        for item in Instruments:
+            value = item.value
+            if value.identifier == blColMidi.instrument_type:
+                value.cls.drawObject(context, col, obj)
+                break
         
 
 class VIEW3D_PT_add_notes_quick(MIDIAniamtorPanel, bpy.types.Panel):
